[Feat] post-detail GET api
